removed redundant component
