+++ src/components/header.js
@@ -1,13 +1,34 @@
 import React from "react"
-import HeaderImage from "./headerImage"
+import { useStaticQuery, graphql } from "gatsby"
+import Img from "gatsby-image"
 
-const Header = () => (
-  <header>
+const Header = () => {
+  const data = useStaticQuery(graphql`
+    query {
+      placeholderImage: file(relativePath: { eq: "ali-sene-new.png" }) {
+        childImageSharp {
+          fluid(maxWidth: 1800) {
+            ...GatsbyImageSharpFluid
+          }
+        }
+      }
+    }
+  `)
 
-      <HeaderImage />
+  return (
 
-  </header>
-)
+    <div
+    style={{
+      display: `block`,
+      maxWidth: `400px`,
+      minWidth: `310px`,
+      margin: `0 auto`,
+    }}
+    >
+      <Img fluid={data.placeholderImage.childImageSharp.fluid} />
+    </div>
 
+  )
+}
 
 export default Header
