--- src/components/headerImage.js
@@ -1,46 +1,0 @@
-import React from "react"
-import { useStaticQuery, graphql } from "gatsby"
-import Img from "gatsby-image"
-
-/*
- * This component is built using `gatsby-image` to automatically serve optimized
- * images with lazy loading and reduced file sizes. The image is loaded using a
- * `useStaticQuery`, which allows us to load the image from directly within this
- * component, rather than having to pass the image data down from pages.
- *
- * For more information, see the docs:
- * - `gatsby-image`: https://gatsby.dev/gatsby-image
- * - `useStaticQuery`: https://www.gatsbyjs.org/docs/use-static-query/
- */
-
-const HeaderImage = () => {
-  const data = useStaticQuery(graphql`
-    query {
-      placeholderImage: file(relativePath: { eq: "ali-sene-new.png" }) {
-        childImageSharp {
-          fluid(maxWidth: 1800) {
-            ...GatsbyImageSharpFluid
-          }
-        }
-      }
-    }
-  `)
-
-  return (
-  <div className="headerImgContainer">
-    <div
-    style={{
-      display: `block`,
-      maxWidth: `400px`,
-      minWidth: `310px`,
-      // minHeight: `150px`,
-      margin: `0 auto`,
-    }}
-    >
-      <Img fluid={data.placeholderImage.childImageSharp.fluid} />
-    </div>
-  </div>
-  )
-}
-
-export default HeaderImage
